requirements.txt
scikit-learn>=1.0.0
mediapipe>=0.8.9
pygame>=2.1.2
numpy>=1.22.0,<1.24.0
pillow>=8.3.0
joblib>=1.1.0
kagglehub>=0.1.5
pandas>=1.3.0
tensorflow==2.12.0
dlib-bin>=19.22.0
face-recognition>=1.3.0 